backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSocket support
websockets==12.0

# Environment variables
python-dotenv==1.0.1

# Existing dependencies (from CLI version)
chromadb>=1.1.1
requests>=2.31.0

# Testing (optional for now)
pytest==7.4.4
pytest-asyncio==0.23.4
httpx>=0.27.0
